+++ src/test/java/lotto/domain/ticket/LottoStoreTest.java
@@ -1,7 +1,6 @@
-package lotto.domain;
+package lotto.domain.ticket;
 
 import lotto.domain.result.win.WinningLotto;
-import lotto.domain.ticket.RealLottoStore;
 import lotto.domain.ticket.ball.LottoBall;
 import lotto.domain.ticket.ball.LottoFactory;
 import lotto.view.dto.WinningLottoRequestDTO;
@@ -14,7 +13,7 @@ import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class RealLottoStoreTest {
+class LottoStoreTest {
 
     @DisplayName("우승 로또 만들기")
     @Test
@@ -27,7 +26,7 @@ class RealLottoStoreTest {
         WinningLottoRequestDTO winningLottoRequestDTO = new WinningLottoRequestDTO("1,2,3,4,5,6", 7);
 
         //when
-        WinningLotto winningLotto = RealLottoStore.makeWinningLotto(winningLottoRequestDTO);
+        WinningLotto winningLotto = LottoStore.makeWinningLotto(winningLottoRequestDTO);
 
         //then
         assertThat(winningLotto).isEqualTo(expectedLotto);
