+++ src/test/java/lotto/domain/ticket/LottoTicketTest.java
@@ -1,6 +1,5 @@
-package lotto.domain;
+package lotto.domain.ticket;
 
-import lotto.domain.ticket.LottoTicket;
 import lotto.domain.ticket.ball.LottoBall;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
