+++ src/test/java/lotto/domain/ticket/ball/LottoBallTest.java
@@ -1,6 +1,5 @@
-package lotto.domain;
+package lotto.domain.ticket.ball;
 
-import lotto.domain.ticket.ball.LottoBall;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
